Reconfigure production build output
Assets should go in the `screen-app` subdirectory and the main file should be called `kiosign_screens.php`, which is the file used to render the screen post type within KioSigns. This then allows that post type to load the app, pass the screen ID and API information into the app and then render it.