Prevent activation to same program
When a program was active and a recalculation occurred with the active programs, there were cases where the program would “re-activate” causing it to switch to the other player and have the transition play again.

This fix ensures that if the old program is equivalent to the current program, no change occurs.